dom/base/test/test_window_keys.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1372371
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1372371</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1372371">Mozilla Bug 1372371</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe></iframe>
</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  /** Test for Bug 1372371 **/
  var list = Object.keys(window);
  is(list.indexOf("WebSocket"), -1, "WebSocket should not be enumerable");
  is(list.indexOf("0"), 0, "Indexed props should be enumerable");
  </script>
</body>
</html>
